+++ DB/dbo/Tables/Sync Tables/_key_constraints_history.sql
@@ -1,8 +1,8 @@
 CREATE TABLE dbo._key_constraints_history (
     _DatabaseID         int           NOT NULL,
-    _ObjectID           int           NOT NULL,
-    _IndexID            int           NOT NULL,
-    _ParentObjectID     int           NOT NULL,
+    _ObjectID           bigint        NOT NULL,
+    _IndexID            bigint        NOT NULL,
+    _ParentObjectID     bigint        NOT NULL,
     --
     _InsertDate         datetime2     NOT NULL,
     _ModifyDate         datetime2     NOT NULL,
