+++ DB/dbo/Tables/Sync Tables/_tables_history.sql
@@ -1,6 +1,6 @@
 CREATE TABLE dbo._tables_history (
     _DatabaseID                        int           NOT NULL,
-    _ObjectID                          int           NOT NULL,
+    _ObjectID                          bigint        NOT NULL,
     --
     _InsertDate                        datetime2     NOT NULL,
     _ModifyDate                        datetime2     NOT NULL,
